delete unreachable/unused code for "update_region_map"
