--- koku/masu/api/region_map.py
@@ -1,35 +1,0 @@
-#
-# Copyright 2018 Red Hat, Inc.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#
-
-"""View endpoint for updating region mapping."""
-
-from django.views.decorators.cache import never_cache
-from rest_framework.decorators import api_view, permission_classes, renderer_classes
-from rest_framework.permissions import AllowAny
-from rest_framework.response import Response
-from rest_framework.settings import api_settings
-
-from masu.util.aws.region_map import update_region_mapping
-
-
-@never_cache
-@api_view(http_method_names=['GET'])
-@permission_classes((AllowAny,))
-@renderer_classes(tuple(api_settings.DEFAULT_RENDERER_CLASSES))
-def update_region_map(request):
-    """Return download file async task ID."""
-    return Response(update_region_mapping())
